+++ pages/mgmIngredient.php
@@ -168,7 +168,19 @@ function reload_data() {
 			$('#fetch_composition').html(data);
 		},
 	});
-
+	
+	$.ajax({ 
+		url: 'views/ingredients/techData.php', 
+		type: 'POST',
+		data: {
+			ingID: "<?=$ing['id']?>",
+		},
+		dataType: 'html',
+		success: function (data) {
+			$('#fetch_tech_data').html(data);
+		},
+	});
+	
 	$.ajax({ 
 		url: 'ingSuppliers.php', 
 		type: 'GET',
@@ -472,46 +484,7 @@ function reload_data() {
 </div>
 
 <div class="tab-pane fade" id="tech_data">
-	<h3>Techical Data</h3>
-	<hr>
-	<table width="100%" border="0">
-		<tr>
-			<td width="20%">Tenacity:</td>
-			<td width="80%" colspan="3"><input name="tenacity" type="text" class="form-control" id="tenacity" value="<?php echo $ing['tenacity']; ?>"/></td>
-		</tr>
-		<tr>
-			<td>Flash Point:</td>
-			<td colspan="3"><input name="flash_point" type="text" class="form-control" id="flash_point" value="<?php echo $ing['flash_point']; ?>"/></td>
-		</tr>
-		<tr>
-			<td>Chemical Name:</td>
-			<td colspan="3"><input name="chemical_name" type="text" class="form-control" id="chemical_name" value="<?php echo $ing['chemical_name']; ?>"/></td>
-		</tr>
-		<tr>
-			<td>Molecular Formula:</td>
-			<td colspan="3">
-            <input name="formula" type="text" class="form-control" id="molecularFormula" value="<?php echo $ing['formula']; ?>">
-			</td>
-		</tr>
-		<tr>
-			<td>Log/P:</td>
-			<td colspan="3"><input name="logp" type="text" class="form-control" id="logp" value="<?php echo $ing['logp']; ?>"/></td>
-		</tr>
-		<tr>
-			<td>Soluble in:</td>
-			<td colspan="3"><input name="soluble" type="text" class="form-control" id="soluble" value="<?php echo $ing['soluble']; ?>"/></td>
-		</tr>
-		<tr>
-			<td>Molecular Weight:</td>
-			<td colspan="3"><input name="molecularWeight" type="text" class="form-control" id="molecularWeight" value="<?php echo $ing['molecularWeight']; ?>"/></td>
-		</tr>
-		<tr>
-			<td>Appearance:</td>
-			<td colspan="3"><input name="appearance" type="text" class="form-control" id="appearance" value="<?php echo $ing['appearance']; ?>"/></td>
-		</tr>
-	</table>
-	<hr />
-	<p><input type="submit" name="save" class="btn btn-info" id="saveTechData" value="Save" /></p>
+	<div id="fetch_tech_data"><div class="loader"></div></div>
 </div>
 
 <div class="tab-pane fade" id="safety_info">
@@ -795,31 +768,6 @@ $(document).ready(function() {
 			success: function (data) {
 				$('#ingMsg').html(data);
 				reload_overview();
-			}
-		});
-	});
-
-
-	$('#tech_data').on('click', '[id*=saveTechData]', function () {
-		$.ajax({ 
-			url: 'update_data.php', 
-			type: 'POST',
-			data: {
-				manage: 'ingredient',
-				tab: 'tech_data',
-				ingID: '<?=$ing['id'];?>',
-				tenacity: $("#tenacity").val(),
-				flash_point: $("#flash_point").val(),
-				chemical_name: $("#chemical_name").val(),
-				formula: $("#formula").val(),
-				logp: $("#logp").val(),
-				soluble: $("#soluble").val(),
-				molecularWeight: $("#molecularWeight").val(),
-				appearance: $("#appearance").val(),
-			},
-			dataType: 'html',
-			success: function (data) {
-				$('#ingMsg').html(data);
 			}
 		});
 	});